requirements.txt
fastapi~=0.65.1
SQLAlchemy~=1.4.15
passlib~=1.7.4
pydantic~=1.8.2
uvicorn~=0.13.4
python-jose[cryptography]
PyQt5